refactor(electron): extract bootstrap into shared module
Both index.js and index.electron.js now require ./src/bootstrap directly
rather than one depending on the other. index.js remains the only
entrypoint that registers OpenFeature.
